Fix hex format string initialization
For some reason FormatHexValue was creating format strings on
demand.  In a recent change, FormatAdjustment started using them.
So if the first 4-digit hex value printed by the program was a
large adjustment, you could end up with a default-formatted adjustment
pretending to be hex.

Now we just create the 4 format strings in the Formatter constructor.